Add same test for no-augmentation, and move them to client
